\name{examples}
\alias{dsfcruesexample}
\alias{dsexample}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{Examples}
\description{
Two self-contained examples, see commented code.
}
\usage{
dsfcruesexample()
dsexample()
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\details{
Please see the code of the examples.
}
\value{
}
\references{Limbourg, P., R. Savic, et al. (2007). Fault Tree Analysis in an Early Design Stage using the Dempster-Shafer Theory of Evidence. European Conference on Safety and Reliability - ESREL 2007, Stavanger, Norway, Taylor and Francis.}
\author{Philipp Limbourg <p.limbourg@uni-due.de>}
\examples{
print("Example Fcrues")
print("See code")
print(dsfcruesexample)
print("Execute example")
dsfcruesexample()
}