\name{ipptoolbox-package}
\alias{ipptoolbox-package}
\alias{ipptoolbox}
\docType{package}
\title{
IPP Toolbox
}
\description{
An R package for uncertainty quantification and propagation in the framework of Dempster-Shafer Theory and imprecise probabilities.

This package was developed in the context of a collaborative research project between EDF R&D (http//rd.edf.com) and the University of Duisburg-Essen, Information Logistics (http//www.uni-due.de/il).
}
\details{
\tabular{ll}{
Package: \tab ipptoolbox\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-11-04\cr
License: \tab GPL version 2 or newer\cr
LazyLoad: \tab yes\cr
}
The package contains a set of methods for conducting uncertainty studies using imprecise probabilities. Please type help(dsfcruesexample) for an example.
}
\author{
Philipp Limbourg <p.limbourg@uni-due.de>

Maintainer: Philipp Limbourg <p.limbourg@uni-due.de>
}
\references{
Ferson, S., V. Kreinovich, et al. (2003). Constructing Probability Boxes and Dempster-Shafer Structures. Sandia Report. Albuquerque, USA, Sandia National Laboratories.

Limbourg, P. (2008). Dependability Modelling under Uncertainty. Berlin, Heidelberg, Springer.
}
\keyword{ package }
\examples{
print("Example Fcrues")
print("See code")
print(dsfcruesexample)
print("Execute example")
dsfcruesexample()
}
